# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing players, tournaments and prizes.",
        "title": "iGaming Platform API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/players/ranking": {
            "get": {
                "description": "Retrieve a ranking list of all players based on their account balances.",
                "tags": [
                    "Players"
                ],
                "summary": "Get global player ranking",
                "responses": {
                    "200": {
                        "description": "Player ranking fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/player.PlayerRanking"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to fetch player ranking",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tournaments/{id}/distribute": {
            "post": {
                "description": "Distribute prizes based on tournament ranking",
                "tags": [
                    "Tournaments"
                ],
                "summary": "Distribute prizes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prizes successfully distributed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "common.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "player.PlayerRanking": {
            "type": "object",
            "properties": {
                "account_balance": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "player_rank": {
                    "type": "integer"
                }
            }
        }
    }
}